mmdetection/efficientDet/others/requirements.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
google-auth==2.6.3
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
kiwisolver==1.3.1
KNN-CUDA @ https://github.com/unlimblue/KNN_CUDA/releases/download/0.2/KNN_CUDA-0.2-py3-none-any.whl
Markdown==3.3.6
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.64
Pillow @ file:///tmp/build/80754af9/pillow_1625649052827/work
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
torch==1.4.0
torchvision==0.5.0
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
webcolors==1.11.1
Werkzeug==2.0.3
zipp==3.6.0
